Scripts for tracking Urban Dictionary's defined words
These scripts are ready to go for tracking the words that Urban
Dictionary bot defines for people. The output are two CSV files
containing word leaderboards for the past seven days and all time. To
keep up with the account, scripts should be run at least once every
other day.